Makefile, .JARs, example, and project file reorganization.
The source is now located under the `src` directory and the provided palette
defintion files are now located under the `palettes` directory.

Added a Makefile. The project can now be compiled by issuing `make`.
`make clean` is also a valid target. Compiled files are now located under the
`build` directory. Compilation also generates `.jar` files for easier usage.

The READMEs and source code have been updated to reflect the fact that the
utilities are built by default as `.jar` files.

I've included a small example which can be compiled with `make example`. It
takes the provided _Vampyr_ game files under the `example` directory and uses
the utilites to decode the contents and put them in the `example/output`
directory. The generated files are also removed with `make clean`.